fix script codegen feature settings [skip ci]
